requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
elastic-transport==8.17.1
elasticsearch==9.0.1
idna==3.10
kombu==5.5.3
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==6.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
